+++ tests/fixtures/models/alias-relationships.php
@@ -2,13 +2,10 @@
 
 namespace App;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Salesman extends Model
 {
-    use HasFactory;
-
     /**
      * The attributes that are mass assignable.
      *
