+++ tests/fixtures/models/all-column-types.php
@@ -2,13 +2,10 @@
 
 namespace App;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class AllType extends Model
 {
-    use HasFactory;
-
     /**
      * The attributes that are mass assignable.
      *
