+++ tests/fixtures/models/certificate-pascal-case-example-laravel8.php
@@ -2,10 +2,13 @@
 
 namespace App;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Certificate extends Model
 {
+    use HasFactory;
+
     /**
      * The attributes that are mass assignable.
      *
