+++ tests/fixtures/models/certificate-type-pascal-case-example.php
@@ -2,13 +2,10 @@
 
 namespace App;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class CertificateType extends Model
 {
-    use HasFactory;
-
     /**
      * The attributes that are mass assignable.
      *
