+++ tests/fixtures/models/custom-models-namespace-laravel8.php
@@ -1,11 +1,15 @@
 <?php
 
-namespace App\Admin;
+namespace App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\SoftDeletes;
 
-class User extends Model
+class Tag extends Model
 {
+    use HasFactory, SoftDeletes;
+
     /**
      * The attributes that are mass assignable.
      *
@@ -13,16 +17,8 @@ class User extends Model
      */
     protected $fillable = [
         'name',
-        'password',
-    ];
-
-    /**
-     * The attributes that should be hidden for serialization.
-     *
-     * @var array
-     */
-    protected $hidden = [
-        'password',
+        'icon',
+        'active',
     ];
 
     /**
@@ -32,5 +28,6 @@ class User extends Model
      */
     protected $casts = [
         'id' => 'integer',
+        'active' => 'boolean',
     ];
 }
