+++ tests/fixtures/models/disable-auto-columns-phpdoc-laravel8.php
@@ -2,10 +2,18 @@
 
 namespace App;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
+/**
+ * @property string $name
+ * @property string $code
+ * @property int $country_id
+ */
 class State extends Model
 {
+    use HasFactory;
+
     /**
      * Indicates if the model should be timestamped.
      *
