+++ tests/fixtures/models/disable-auto-columns-phpdoc.php
@@ -2,7 +2,6 @@
 
 namespace App;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 /**
@@ -12,8 +11,6 @@ use Illuminate\Database\Eloquent\Model;
  */
 class State extends Model
 {
-    use HasFactory;
-
     /**
      * Indicates if the model should be timestamped.
      *
