+++ tests/fixtures/models/disable-auto-columns.php
@@ -2,13 +2,10 @@
 
 namespace App;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class State extends Model
 {
-    use HasFactory;
-
     /**
      * Indicates if the model should be timestamped.
      *
