+++ tests/fixtures/models/model-configured-laravel8.php
@@ -1,13 +1,13 @@
 <?php
 
-namespace App;
+namespace Some\App\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
 
 class Comment extends Model
 {
-    use SoftDeletes;
+    use HasFactory;
 
     /**
      * The attributes that are mass assignable.
@@ -16,6 +16,7 @@ class Comment extends Model
      */
     protected $fillable = [
         'post_id',
+        'author_id',
     ];
 
     /**
@@ -26,11 +27,17 @@ class Comment extends Model
     protected $casts = [
         'id' => 'integer',
         'post_id' => 'integer',
+        'author_id' => 'integer',
     ];
 
 
     public function post()
     {
-        return $this->belongsTo(\App\Post::class);
+        return $this->belongsTo(\Some\App\Models\Post::class);
+    }
+
+    public function author()
+    {
+        return $this->belongsTo(\Some\App\Models\User::class);
     }
 }
