+++ tests/fixtures/models/model-configured.php
@@ -2,13 +2,10 @@
 
 namespace Some\App\Models;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Comment extends Model
 {
-    use HasFactory;
-
     /**
      * The attributes that are mass assignable.
      *
