+++ tests/fixtures/models/model-guarded.php
@@ -2,13 +2,10 @@
 
 namespace App;
 
-use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class Comment extends Model
 {
-    use HasFactory;
-
     /**
      * The attributes that aren't mass assignable.
      *
