+++ tests/fixtures/models/post-polymorphic-relationship-laravel8.php
@@ -2,19 +2,21 @@
 
 namespace App;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
 
-class Comment extends Model
+class Post extends Model
 {
-    use SoftDeletes;
+    use HasFactory;
 
     /**
      * The attributes that are mass assignable.
      *
      * @var array
      */
-    protected $fillable = [];
+    protected $fillable = [
+        'name',
+    ];
 
     /**
      * The attributes that should be cast to native types.
@@ -24,4 +26,10 @@ class Comment extends Model
     protected $casts = [
         'id' => 'integer',
     ];
+
+
+    public function images()
+    {
+        return $this->morphMany(\App\Image::class, 'imageable');
+    }
 }
