+++ tests/fixtures/models/relationships-laravel8.php
@@ -2,19 +2,12 @@
 
 namespace App;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\SoftDeletes;
 
-/**
- * @property int $id
- * @property int $post_id
- * @property \Carbon\Carbon $deleted_at
- * @property \Carbon\Carbon $created_at
- * @property \Carbon\Carbon $updated_at
- */
 class Comment extends Model
 {
-    use SoftDeletes;
+    use HasFactory;
 
     /**
      * The attributes that are mass assignable.
@@ -23,6 +16,7 @@ class Comment extends Model
      */
     protected $fillable = [
         'post_id',
+        'author_id',
     ];
 
     /**
@@ -33,14 +27,17 @@ class Comment extends Model
     protected $casts = [
         'id' => 'integer',
         'post_id' => 'integer',
+        'author_id' => 'integer',
     ];
 
 
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
-     */
     public function post()
     {
         return $this->belongsTo(\App\Post::class);
+    }
+
+    public function author()
+    {
+        return $this->belongsTo(\App\User::class);
     }
 }
