+++ tests/fixtures/models/relationships-phpdoc-laravel8.php
@@ -2,18 +2,20 @@
 
 namespace App;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 /**
  * @property int $id
  * @property int $post_id
  * @property int $author_id
- * @property int $ccid
  * @property \Carbon\Carbon $created_at
  * @property \Carbon\Carbon $updated_at
  */
 class Comment extends Model
 {
+    use HasFactory;
+
     /**
      * The attributes that are mass assignable.
      *
@@ -22,7 +24,6 @@ class Comment extends Model
     protected $fillable = [
         'post_id',
         'author_id',
-        'ccid',
     ];
 
     /**
@@ -34,7 +35,6 @@ class Comment extends Model
         'id' => 'integer',
         'post_id' => 'integer',
         'author_id' => 'integer',
-        'ccid' => 'integer',
     ];
 
 
@@ -52,13 +52,5 @@ class Comment extends Model
     public function author()
     {
         return $this->belongsTo(\App\User::class);
-    }
-
-    /**
-     * @return \Illuminate\Database\Eloquent\Relations\BelongsTo
-     */
-    public function country()
-    {
-        return $this->belongsTo(\App\Country::class, 'ccid', 'code');
     }
 }
